Support using a float as the retry_interval.

+++ lib/que/job.rb
@@ -37,7 +37,7 @@ module Que
         UPDATE public.que_jobs
 
         SET error_count          = error_count + 1,
-            run_at               = now() + $1::bigint * '1 second'::interval,
+            run_at               = now() + $1::float * '1 second'::interval,
             last_error_message   = $2::text,
             last_error_backtrace = $3::text
